chore: downgrade to dokka 1.9.20
With dokka 2.0.0-Beta, the Gradle Plugin stopped to sync. The error was caused by: java.lang.NoClassDefFoundError: dev/tonholo/s2c/error/ExitProgramException.

I'm unsure why that happened, it appears that dokka was removing classes from the jar file generated by the KMP.

If dokka is applied only in the Gradle Plugin, the error doesn't happen, however when applied in the KMP library, it does.